backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
yfinance>=0.2.28
pandas>=1.5.0
numpy>=1.21.0
scikit-learn
python-dotenv
google-generativeai
requests